third_party/WebKit/LayoutTests/dom/text/previous-element-sibling.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>

description('This tests the NonDocumentTypeChildNode interface previousElementSibling property on a Text node.');

var element = document.createElement('div');
var a = element.appendChild(new Text('a'));
shouldBeNull('a.previousElementSibling');
var b = element.insertBefore(document.createElement('b'), a);
shouldBe('a.previousElementSibling', 'b');
var c = element.insertBefore(new Comment('c'), a);
shouldBe('a.previousElementSibling', 'b');

</script>
