third_party/blink/web_tests/platform/generic/dom/text/previous-element-sibling-expected.txt

This tests the NonDocumentTypeChildNode interface previousElementSibling property on a Text node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS a.previousElementSibling is null
PASS a.previousElementSibling is b
PASS a.previousElementSibling is b
PASS successfullyParsed is true

TEST COMPLETE
